mandala flat removal

+++ Xbus/XbusPacket.h
@@ -4,11 +4,10 @@
 
 namespace xbus {
 
-typedef uint8_t pid_t;
+typedef uint16_t pid_t;
 
 enum {
     size_packet_max = 1024,
 };
 
-
-} // namespace
+} // namespace xbus
